Removed wrap_content from TextViews in Dashboard
This is an first patch in series to improve responsiveness and performance of app.
According to this:
https://medium.com/@nicolasbridoux/chasing-frame-drops-on-android-with-gpu-rendering-inspection-c623240315b9
Very slow on measure may be caused by frequently changed values on TextViews which have wrap_content
in CoordinateLayout.
Indeed, this helps. I've removed wrap_content and used statically designed size, but i feel this is
more like a workaround or hack than regular fix.